Reorganized packages to a domain-oriented approach
The packages are owner, pet, vet and user. We chose to keep pet out of owner because it is the central element of the application. However, visit is included in pet because a visit heavily depends on the pet that makes it. These were also the same packages that were used to organized the JSP code.